+++ app/src/main/java/com/example/FocusApp/screens/LandingScreen.kt
@@ -1,6 +1,5 @@
-package com.example.FocusApp
+package com.example.FocusApp.screens
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Text
@@ -10,7 +9,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import kotlinx.coroutines.delay
 
 /**
