+++ app/src/main/java/com/example/FocusApp/screens/StatsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.FocusApp
+package com.example.FocusApp.screens
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
