+++ app/src/main/java/com/example/FocusApp/screens/TaskGenerator.kt
@@ -1,4 +1,4 @@
-package com.example.FocusApp
+package com.example.FocusApp.screens
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
